flask_blog/requirements.txt
alembic==1.8.1
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.12.0
importlib-resources==5.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-enum==1.5.1
mimesis==6.0.0
packaging==21.3
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.39
Werkzeug==2.1.2
zipp==3.8.0
